+++ worker/include/PayloadChannel/PayloadChannelSocket.hpp
@@ -84,7 +84,7 @@ namespace PayloadChannel
 		void Send(json& jsonMessage);
 
 	private:
-		void SendImpl(const void* nsPayload, size_t nsPayloadLen);
+		void SendImpl(const void* nsPayload, uint32_t nsPayloadLen);
 
 		/* Pure virtual methods inherited from ConsumerSocket::Listener. */
 	public:
